+++ examples/apps/web-todomvc-ace/src/app.js
@@ -1,0 +1,9 @@
+import { html } from "@inglorious/web"
+
+export const app = {
+  render(api) {
+    return html`${api.render("form")}${api.render("list")}${api.render(
+      "footer",
+    )}`
+  },
+}