+++ examples/apps/web-select/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web-select
 
+## 0.2.17
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@4.1.1
+
 ## 0.2.16
 
 ### Patch Changes